Окончательно разбивает на мастер презентер и подчиненные
