--- src/presenter/header-presenter.js
@@ -1,56 +1,0 @@
-import { render, RenderPosition } from '../framework/render.js';
-import TripInfoView from '../view/trip-info-view/trip-info-view.js';
-import FilterContentView from '../view/filter-list-view/filter-list-view.js';
-import SortListView from '../view/sort-list-view/sort-list-view.js';
-
-export default class HeaderPresenter {
-  #tripInfoContainer = null;
-  #filtersListContainer = null;
-  #tripInfoView = new TripInfoView();
-  #filterListView = new FilterContentView({isViewList: true});
-  #filterItemView = null;
-  #filtersModel = null;
-  #sortListComponent = new SortListView();
-  #listContainer = null;
-
-  constructor({ tripInfoContainer, filtersListContainer, filtersModel, listContainer }) {
-    this.#tripInfoContainer = tripInfoContainer;
-    this.#filtersListContainer = filtersListContainer;
-    this.#filtersModel = filtersModel;
-    this.#listContainer = listContainer;
-  }
-
-  init() {
-    this.#renderTripInfo();
-    this.#renderAllFilters();
-    this.#renderSortList();
-  }
-
-  #renderTripInfo() {
-    render(this.#tripInfoView, this.#tripInfoContainer, RenderPosition.AFTERBEGIN);
-  }
-
-  #renderAllFilters() {
-    render(this.#filterListView, this.#filtersListContainer);
-
-    const allFilters = [...this.#filtersModel.allFilters];
-    allFilters.forEach((filter) => {
-      this.#renderFilter(filter);
-    });
-  }
-
-  #renderFilter(filter) {
-    const { id, name, value, checked } = filter;
-    this.#filterItemView = new FilterContentView({
-      id,
-      name,
-      value,
-      checked,
-      isViewList: false});
-    render(this.#filterItemView, this.#filterListView.element);
-  }
-
-  #renderSortList() {
-    render(this.#sortListComponent, this.#listContainer);
-  }
-}
