+++ src/view/waypoint-list-view/waypoint-list-view-template.js
@@ -1,0 +1,2 @@
+export const waypointItemTemplate = '<ul class="trip-events__list"></ul>';
+
